Backed out merge changeset: 613e4743ed5b
Backed out merge revision to its second parent (b26a7ba9d6fe)

+++ src/uti_phgrm/GpGpu/GpGpu_InterfaceCorrel.cpp
@@ -49,7 +49,7 @@ uint GpGpuInterfaceCorrel::InitCorrelJob(int Zmin, int Zmax)
 }
 
 /// \brief Initialisation des parametres constants
-void GpGpuInterfaceCorrel::SetParameter(int nbLayer , ushort2 dRVig , uint2 dimImg, float mAhEpsilon, uint samplingZ, int uvINTDef )
+void GpGpuInterfaceCorrel::SetParameter(int nbLayer , uint2 dRVig , uint2 dimImg, float mAhEpsilon, uint samplingZ, int uvINTDef )
 {
 
     if(!copyInvParam || _param[0].invPC.nbImages != (uint)nbLayer || _param[1].invPC.nbImages != (uint)nbLayer)
@@ -156,7 +156,6 @@ void GpGpuInterfaceCorrel::SetTexturesAreLoaded(bool load)
 void GpGpuInterfaceCorrel::CorrelationGpGpu(ushort idBuf,const int s )
 {
     LaunchKernelCorrelation(s, *(GetStream(s)),_param[idBuf], _data2Cor);
-    //LaunchKernelCorrelationZ(s,_param[idBuf],_data2Cor);
 }
 
 void GpGpuInterfaceCorrel::MultiCorrelationGpGpu(ushort idBuf, const int s)
